splits Analytics record generation from Service
